+++ src/main/java/org/satocash/applet/Satocash.java
@@ -59,11 +59,13 @@ public class Satocash extends javacard.framework.Applet {
      * APPLET VERSION:   changes with no impact on compatibility of the client
      *
      * 0.1-0.1: (WIP) initial version
+     * 0.1-0.2: (WIP) support import authentikey during personalization. This key can be shared by multiple devices for privacy.
+     * 0.1-0.3: (WIP) Add amount limits for PIN-less payments.
      */
     private final static byte PROTOCOL_MAJOR_VERSION = (byte) 0; 
     private final static byte PROTOCOL_MINOR_VERSION = (byte) 1;
     private final static byte APPLET_MAJOR_VERSION = (byte) 0;
-    private final static byte APPLET_MINOR_VERSION = (byte) 1;   
+    private final static byte APPLET_MINOR_VERSION = (byte) 3;
 
     // Maximum size for the extended APDU buffer 
     private final static short EXT_APDU_BUFFER_SIZE = (short) 320;
@@ -138,6 +140,7 @@ public class Satocash extends javacard.framework.Applet {
     private final static byte INS_EXPORT_PKI_CERTIFICATE = (byte) 0x93;
     private final static byte INS_SIGN_PKI_CSR = (byte) 0x94;
     private final static byte INS_EXPORT_PKI_PUBKEY = (byte) 0x98;
+    private final static byte INS_IMPORT_PKI_NDEF_AUTHENTIKEY = (byte) 0x9B;
     private final static byte INS_LOCK_PKI = (byte) 0x99;
     private final static byte INS_CHALLENGE_RESPONSE_PKI= (byte) 0x9A;
     
@@ -229,13 +232,11 @@ public class Satocash extends javacard.framework.Applet {
     private final static short SW_DEBUG_FLAG = (short) 0x9FFF;
 
     /** Satocash errors */
-    private final static short SW_OBJECT_ALREADY_PRESENT = (short) 0x9C60;
-
+    //private final static short SW_OBJECT_ALREADY_PRESENT = (short) 0x9C60;
     /** Math error */
-    private final static short SW_MATH_OVERFLOW = (short) 0x9999;
-
+    private final static short SW_MATH_OVERFLOW = (short) 0x9C61;
     /** Multiple units error */
-    private final static short SW_MULTIPLE_UNITS = (short) 0x9990;
+    private final static short SW_MULTIPLE_UNITS = (short) 0x9C62;
 
 
     // KeyBlob Encoding in Key Blobs
@@ -317,6 +318,7 @@ public class Satocash extends javacard.framework.Applet {
      *  BIP32 Hierarchical Deterministic Wallet  *
      *********************************************/
 
+    private static final short SIZE_ECPRIVKEY= (short)32;
     private static final short BIP32_KEY_SIZE= 32; // size of extended key and chain code is 256 bits
 
 
@@ -343,7 +345,7 @@ public class Satocash extends javacard.framework.Applet {
     // Proofs table
     private byte[] proofs;
     private static final short MAX_NB_PROOFS = 128; // todo: configurable in constructor
-    private static final byte MAX_PROOF_EXPONENT = 24; // 2^24 * 2^7 == 2^31
+    private static final byte MAX_PROOF_EXPONENT = 31;
     private static final byte PROOF_OBJECT_SIZE = 68;
     private static final byte PROOF_OFFSET_STATE = 0; // 1 byte
     private static final byte PROOF_OFFSET_KEYSET_INDEX = 1; // 1 byte
@@ -383,7 +385,7 @@ public class Satocash extends javacard.framework.Applet {
     private static final byte UNIT_MSAT = 2;
     private static final byte UNIT_USD = 3;
     private static final byte UNIT_EUR = 4;
-    private static final short UNITS_LENGTH = 4;
+    private static final short NB_UNITS = 4;
     private static final byte AMOUNT_NULL = (byte) 0xff;
 
     // metadata_type constants
@@ -401,7 +403,9 @@ public class Satocash extends javacard.framework.Applet {
     private static final byte[] CST_SC = {'s','c','_','k','e','y', 's','c','_','m','a','c'};
     private boolean needs_secure_channel= true;
     private boolean initialized_secure_channel= false;
-    private ECPrivateKey sc_ephemeralkey; 
+    //private ECPrivateKey sc_ephemeralkey;
+    private ECPrivateKey ephemeral_privkey;
+    private boolean ephemeral_privkey_transient = false;
     private AESKey sc_sessionkey;
     private Cipher sc_aes128_cbc;
     private byte[] sc_buffer;
@@ -478,9 +482,9 @@ public class Satocash extends javacard.framework.Applet {
         // default PIN policy: require pin to change state or make payment
         pin_policy = PIN_POLICY_MASK_CHANGE_STATE | PIN_POLICY_MASK_MAKE_PAYMENT;
         // Threshold for all units requires allocating (how_many_units) * (bytes for threshold) bytes
-        pin_policy_amount_threshold = new byte[UNITS_LENGTH * PIN_POLICY_AMOUNT_THRESHOLD_SIZE];
+        pin_policy_amount_threshold = new byte[NB_UNITS * PIN_POLICY_AMOUNT_THRESHOLD_SIZE];
         // Cumulative pinless amount tracking for all units
-        pin_policy_cumulative_amount = new byte[UNITS_LENGTH * PIN_POLICY_AMOUNT_THRESHOLD_SIZE];
+        pin_policy_cumulative_amount = new byte[NB_UNITS * PIN_POLICY_AMOUNT_THRESHOLD_SIZE];
 
         // Temporary working arrays
         try {
@@ -539,14 +543,16 @@ public class Satocash extends javacard.framework.Applet {
 
         try {
             // Put the EC key in RAM if we can.
-            sc_ephemeralkey= (ECPrivateKey) KeyBuilder.buildKey(TYPE_EC_FP_PRIVATE_TRANSIENT_DESELECT, LENGTH_EC_FP_256, false);
+            ephemeral_privkey= (ECPrivateKey) KeyBuilder.buildKey(TYPE_EC_FP_PRIVATE_TRANSIENT_DESELECT, LENGTH_EC_FP_256, false);
+            ephemeral_privkey_transient = true;
         } catch (CryptoException e) {
             try {
                 // This uses a bit more RAM, but at least it isn't using flash.
-                sc_ephemeralkey= (ECPrivateKey) KeyBuilder.buildKey(TYPE_EC_FP_PRIVATE_TRANSIENT_RESET, LENGTH_EC_FP_256, false);
+                ephemeral_privkey= (ECPrivateKey) KeyBuilder.buildKey(TYPE_EC_FP_PRIVATE_TRANSIENT_RESET, LENGTH_EC_FP_256, false);
+                ephemeral_privkey_transient = true;
             } catch (CryptoException x) {
                 // Last option as it will wear out the flash eventually
-                sc_ephemeralkey= (ECPrivateKey) KeyBuilder.buildKey(KeyBuilder.TYPE_EC_FP_PRIVATE, LENGTH_EC_FP_256, false);
+                ephemeral_privkey= (ECPrivateKey) KeyBuilder.buildKey(KeyBuilder.TYPE_EC_FP_PRIVATE, LENGTH_EC_FP_256, false);
             }
         }
         sc_aes128_cbc= Cipher.getInstance(Cipher.ALG_AES_BLOCK_128_CBC_NOPAD, false);
@@ -787,6 +793,9 @@ public class Satocash extends javacard.framework.Applet {
             case INS_SIGN_PKI_CSR:
                 sizeout= sign_PKI_CSR(apdu, buffer);
                 break;
+            case INS_IMPORT_PKI_NDEF_AUTHENTIKEY:
+                sizeout= import_PKI_ndef_authentikey(apdu, buffer);
+                break;
             case INS_LOCK_PKI:
                 sizeout= lock_PKI(apdu, buffer);
                 break;
@@ -952,9 +961,6 @@ public class Satocash extends javacard.framework.Applet {
             base+=(short)2;
             bytesLeft-=(short)2;
         }
-        
-        // bip32
-//        Secp256k1.setCommonCurveParameters(bip32_extendedkey);
 
         om_nextid= (short)0;
         setupDone = true;
@@ -980,15 +986,16 @@ public class Satocash extends javacard.framework.Applet {
         // logs
         // currently, we do NOT erase logs, but we add an entry for the reset
         logger.createLog(INS_RESET_TO_FACTORY, (short)-1, (short)-1, (short)0x0000 );
-        
+
         // reset proofs, mints & keysets
         Util.arrayFillNonAtomic(proofs, (short)0, (short)proofs.length, (byte)0);
         Util.arrayFillNonAtomic(keysets, (short)0, (short)keysets.length, (byte)0);
         Util.arrayFillNonAtomic(mints, (short)0, (short)mints.length, (byte)0);
 
-        // reset pin_policy_amount_threshold
-        Util.arrayFillNonAtomic(pin_policy_amount_threshold, (short)0, (short)(UNITS_LENGTH*PIN_POLICY_AMOUNT_THRESHOLD_SIZE), (byte)0x00);
-        Util.arrayFillNonAtomic(pin_policy_cumulative_amount, (short)0, (short)(UNITS_LENGTH*PIN_POLICY_AMOUNT_THRESHOLD_SIZE), (byte)0x00);
+        // reset pin_policy & pin_policy_amount_threshold
+        pin_policy = PIN_POLICY_MASK_CHANGE_STATE | PIN_POLICY_MASK_MAKE_PAYMENT;
+        Util.arrayFillNonAtomic(pin_policy_amount_threshold, (short)0, (short)(NB_UNITS *PIN_POLICY_AMOUNT_THRESHOLD_SIZE), (byte)0x00);
+        Util.arrayFillNonAtomic(pin_policy_cumulative_amount, (short)0, (short)(NB_UNITS *PIN_POLICY_AMOUNT_THRESHOLD_SIZE), (byte)0x00);
 
         // reset export list
         short index;
@@ -1042,7 +1049,10 @@ public class Satocash extends javacard.framework.Applet {
      *            pin_policy(1b) | RFU(1b) |
      *            MAX_NB_MINT (1b) | NB_USED_MINT(1b) |
      *            MAX_NB_KEYSET(1b) | NB_USED_KEYSET(1b)  |
-     *            MAX_NB_PROOFS(2b) | NB_PROOFS_UNSPENT(2b) | NB_PROOFS_SPENT(2b)
+     *            MAX_NB_PROOFS(2b) | NB_PROOFS_UNSPENT(2b) | NB_PROOFS_SPENT(2b) |
+     *            RFU(6b) |
+     *            size(NB_UNITS*PIN_POLICY_AMOUNT_THRESHOLD_SIZE) | pin_policy_amount_threshold |
+     *            size(NB_UNITS*PIN_POLICY_AMOUNT_THRESHOLD_SIZE) | pin_policy_cumulative_amount
      *          ]
      *
      *  Exceptions: (none)
@@ -1101,7 +1111,24 @@ public class Satocash extends javacard.framework.Applet {
         pos+=2;
         Util.setShort(buffer, pos, NB_PROOFS_SPENT);
         pos+=2;
-        // todo max amount allowed without PIN (pin_policy_amount_threshold)
+
+        // RFU (6b)
+        Util.arrayFillNonAtomic(buffer, pos, (short)6, (byte)0x00);
+        pos+=6;
+
+        // max amount allowed without PIN (pin_policy_amount_threshold)
+        // format [size(NB_UNITS*PIN_POLICY_AMOUNT_THRESHOLD_SIZE) | pin_policy_amount_threshold]
+        byte amounts_size = (byte) (NB_UNITS *PIN_POLICY_AMOUNT_THRESHOLD_SIZE);
+        buffer[pos++] = amounts_size;
+        Util.arrayCopyNonAtomic(pin_policy_amount_threshold, (short)0, buffer, pos, amounts_size);
+        pos+=amounts_size;
+
+        // amount already spent without PIN (pin_policy_cumulative_amount)
+        // format [size(NB_UNITS*PIN_POLICY_AMOUNT_THRESHOLD_SIZE) | pin_policy_cumulative_amount]
+        buffer[pos++] = amounts_size;
+        Util.arrayCopyNonAtomic(pin_policy_cumulative_amount, (short)0, buffer, pos, amounts_size);
+        pos+=amounts_size;
+
         return pos;
     }
 
@@ -1514,7 +1541,7 @@ public class Satocash extends javacard.framework.Applet {
      *
      *  return: [proof_index(2b) | proof_state(1b) | keyset_index(1b) | amount_exponent(1b) | unblinded_key(33b) | secret(32b)]
      *
-     *  exceptions (OP_INIT): 9C06 SW_UNAUTHORIZED, 9C11 SW_INCORRECT_P2, 6700 SW_WRONG_LENGTH, 9C0F SW_INVALID_PARAMETER,
+     *  exceptions (OP_INIT): 9C06 SW_UNAUTHORIZED, 9C11 SW_INCORRECT_P2, 6700 SW_WRONG_LENGTH, 9C0F SW_INVALID_PARAMETER, 9C61 SW_MATH_OVERFLOW, 9C62 SW_MULTIPLE_UNITS,
      *  exceptions (OP_PROCESS): 9C06 SW_UNAUTHORIZED, 9C11 SW_INCORRECT_P2, 9C13 SW_INCORRECT_INITIALIZATION,
      */
     private short satocashExportProofs(APDU apdu, byte[] buffer) {
@@ -1557,62 +1584,81 @@ public class Satocash extends javacard.framework.Applet {
                     if (index_proof<0 || index_proof>= MAX_NB_PROOFS)
                         ISOException.throwIt(SW_INVALID_PARAMETER);
 
-
-                    // unit from the first proof is the reference for future proofs
+                    // check unit consistency
+                    byte index_keyset = proofs[(short)(index_proof * PROOF_OBJECT_SIZE + PROOF_OFFSET_KEYSET_INDEX)];
                     if (index_in == 0) {
-                        byte index_selected_keyset = proofs[(short)(index_proof * PROOF_OBJECT_SIZE + PROOF_OFFSET_KEYSET_INDEX)];
-                        selected_unit = keysets[(short)(index_selected_keyset * KEYSET_OBJECT_SIZE + KEYSET_OFFSET_UNIT)];
-                        if (selected_unit == UNIT_NONE) {
+                        // unit from the first proof is the reference for future proofs
+                        selected_unit = keysets[(short)(index_keyset * KEYSET_OBJECT_SIZE + KEYSET_OFFSET_UNIT)];
+                        // check that selected unit is valid
+                        if (selected_unit <= UNIT_NONE) {
                             ISOException.throwIt(SW_INVALID_PARAMETER);
                         }
-                    }
-
-                    // if the unit is not the same, throw an error
-                    byte index_keyset = proofs[(short)(index_proof * PROOF_OBJECT_SIZE + PROOF_OFFSET_KEYSET_INDEX)];
-                    byte unit = keysets[(short)(index_keyset * KEYSET_OBJECT_SIZE + KEYSET_OFFSET_UNIT)];
-                    if (unit != selected_unit) {
-                        ISOException.throwIt(SW_MULTIPLE_UNITS);
-                    }
-
-                    // Get the amount of the proof and accumulate it into `tmpBuffer`
-                    // Amount exponent needs to be converted to a number
-                    Util.arrayFillNonAtomic(tmpBuffer2, (short)0, PIN_POLICY_AMOUNT_THRESHOLD_SIZE, (byte)0x00);
-                    byte exponent = proofs[(short)(index_proof * PROOF_OBJECT_SIZE + PROOF_OFFSET_AMOUNT_EXPONENT)];
-                    if (exponent != 0xFF) {
-                        // exponent can't be > 24
-                        byte div = (byte)(exponent >> 3);
-                        // div can't be > 3
-                        byte rem = (byte)(exponent - 8*div);
-                        tmpBuffer2[3-div] = (byte)(1 << rem);
-                        if (Biginteger.add_carry(tmpBuffer, (short)0, tmpBuffer2, (short)0, (short)4)) {
-                            ISOException.throwIt(SW_MATH_OVERFLOW);
+                        // if selected unit is not in the range for which amount threshold is defined, PIN is required
+                        if (selected_unit > NB_UNITS) {
+                            if ((pin_policy & PIN_POLICY_MASK_MAKE_PAYMENT) == PIN_POLICY_MASK_MAKE_PAYMENT){
+                                if (!pin.isValidated()) {
+                                    ISOException.throwIt(SW_UNAUTHORIZED);
+                                }
+                            }
+                        }
+                    } else {
+                        // if the unit is not the same as reference, throw an error
+                        byte unit = keysets[(short)(index_keyset * KEYSET_OBJECT_SIZE + KEYSET_OFFSET_UNIT)];
+                        if (unit != selected_unit) {
+                            ISOException.throwIt(SW_MULTIPLE_UNITS);
                         }
                     }
+
+                    // check PIN policy
+                    if ((pin_policy & PIN_POLICY_MASK_MAKE_PAYMENT) == PIN_POLICY_MASK_MAKE_PAYMENT){
+
+                        // compute total proof amount
+                        if (!pin.isValidated()) {
+
+                            // Get the amount of the proof and accumulate it into `tmpBuffer`
+                            // Amount exponent needs to be converted to a number
+                            Util.arrayFillNonAtomic(tmpBuffer2, (short)0, PIN_POLICY_AMOUNT_THRESHOLD_SIZE, (byte)0x00);
+                            byte exponent = proofs[(short)(index_proof * PROOF_OBJECT_SIZE + PROOF_OFFSET_AMOUNT_EXPONENT)];
+                            if (exponent != AMOUNT_NULL) {
+                                // exponent can't be > 31
+                                byte div = (byte)(exponent >> 3); // div can't be > 3
+                                byte rem = (byte)(exponent & 0x07); // rem can't be > 7
+                                tmpBuffer2[3-div] = (byte)(1 << rem);
+                                if (Biginteger.add_carry(tmpBuffer, (short)0, tmpBuffer2, (short)0, PIN_POLICY_AMOUNT_THRESHOLD_SIZE)) {
+                                    ISOException.throwIt(SW_MATH_OVERFLOW);
+                                }
+                            }
+                        }
+                    }
+
                     // save index_proof in state
                     proof_export_list[index_out++] = index_proof;
                 }
 
-                // if PIN policy requires it, check that PIN has been entered previously
+                // if PIN policy requires it, check if amount threshold is reached
                 if ((pin_policy & PIN_POLICY_MASK_MAKE_PAYMENT) == PIN_POLICY_MASK_MAKE_PAYMENT) {
-                    
-                    // Get the current cumulative amount for this unit
-                    short cumulative_offset = (short)((selected_unit-1) * PIN_POLICY_AMOUNT_THRESHOLD_SIZE);
-                    
-                    // Add the current transaction amount to cumulative amount
-                    if (Biginteger.add_carry(pin_policy_cumulative_amount, cumulative_offset, tmpBuffer, (short)0, PIN_POLICY_AMOUNT_THRESHOLD_SIZE)) {
-                        ISOException.throwIt(SW_MATH_OVERFLOW);
-                    }
-                    
-                    // Check if cumulative amount exceeds threshold and PIN is not validated
-                    if (Biginteger.lessThan(pin_policy_amount_threshold, cumulative_offset, pin_policy_cumulative_amount, cumulative_offset, PIN_POLICY_AMOUNT_THRESHOLD_SIZE)) { 
-                        if (!pin.isValidated()) {
-                            ISOException.throwIt(SW_UNAUTHORIZED);
+
+                    if (!pin.isValidated()) {
+                        // Get the current cumulative amount for this unit
+                        short cumulative_offset = (short) ((selected_unit - 1) * PIN_POLICY_AMOUNT_THRESHOLD_SIZE);
+
+                        // Add the cumulative amount to current transaction amount
+                        if (Biginteger.add_carry(tmpBuffer, (short) 0, pin_policy_cumulative_amount, cumulative_offset, PIN_POLICY_AMOUNT_THRESHOLD_SIZE)) {
+                            ISOException.throwIt(SW_MATH_OVERFLOW);
                         }
-                        
-                        // If the PIN is validated, reset the cumulative amount
+
+                        // Check if new total cumulative amount exceeds threshold
+                        if (Biginteger.lessThan(pin_policy_amount_threshold, cumulative_offset, tmpBuffer, (short) 0, PIN_POLICY_AMOUNT_THRESHOLD_SIZE)) {
+                            ISOException.throwIt(SW_UNAUTHORIZED);
+                        } else {
+                            // update pin_policy_cumulative_amount
+                            Util.arrayCopyNonAtomic(tmpBuffer, (short) 0, pin_policy_cumulative_amount, cumulative_offset, PIN_POLICY_AMOUNT_THRESHOLD_SIZE);
+                        }
+
+                    } else {
+                        // If the PIN is validated, reset the pinless cumulative amount
                         resetCumulativeAmount(selected_unit);
                     }
-                    
 
                 }
 
@@ -1816,7 +1862,7 @@ public class Satocash extends javacard.framework.Applet {
      ****************************************/
 
     /**
-     * Reset cumulative pinless amounts for all units to zero
+     * Reset cumulative pinless amounts for given unit to zero
      * This should be called when PIN is successfully validated
      */
     private void resetCumulativeAmount(short unit) {
@@ -2196,6 +2242,8 @@ public class Satocash extends javacard.framework.Applet {
      *  p2: RFU (set specific permission policies for NFC interface)
      *  data: (none)
      *  return: (none)
+     *
+     *  exceptions: 9C06 SW_UNAUTHORIZED, 9C10 SW_INCORRECT_P1, 9C48 SW_NFC_DISABLED, 9C49 SW_NFC_BLOCKED
      */
     private short setNfcPolicy(APDU apdu, byte[] buffer){
         // check that PIN has been entered previously
@@ -2238,11 +2286,15 @@ public class Satocash extends javacard.framework.Applet {
      *  p2: RFU
      *  data: (none)
      *  return: (none)
+     *
+     *  exceptions: 9C06 SW_UNAUTHORIZED,
      */
     private short setPinPolicy(APDU apdu, byte[] buffer){
         // check that PIN has been entered previously
         if (!pin.isValidated())
             ISOException.throwIt(SW_UNAUTHORIZED);
+
+        // TODO: check 2FA if enabled
 
         // get new PIN policy from P1
         pin_policy = buffer[ISO7816.OFFSET_P1];
@@ -2258,20 +2310,24 @@ public class Satocash extends javacard.framework.Applet {
      * PIN must be validated to call this function.
      *
      *  ins: 0x3B
-     *  p1: RFU
-     *  p2: unit(1b)
+     *  p1: unit(1b)
+     *  p2: RFU
      *  data: [amount(4b)]
      *  return: (none)
+     *
+     *  exceptions: 9C06 SW_UNAUTHORIZED, 9C10 SW_INCORRECT_P1, 6700 SW_WRONG_LENGTH
      */
     private short setPinlessAmount(APDU apdu, byte[] buffer){
         // check that PIN has been entered previously
         if (!pin.isValidated())
             ISOException.throwIt(SW_UNAUTHORIZED);
 
-        byte unit = buffer[ISO7816.OFFSET_P2];
+        // TODO: check 2FA if enabled
+
+        byte unit = buffer[ISO7816.OFFSET_P1];
         // Make sure the unit is valid
-        if (unit < UNIT_SAT || unit > UNIT_EUR) {
-            ISOException.throwIt(SW_INCORRECT_P2);
+        if (unit < UNIT_SAT || unit > NB_UNITS) {
+            ISOException.throwIt(SW_INCORRECT_P1);
         }
 
         short bytesLeft = Util.makeShort((byte) 0x00, buffer[ISO7816.OFFSET_LC]);
@@ -2283,12 +2339,8 @@ public class Satocash extends javacard.framework.Applet {
         // copy array
         Util.arrayCopy(buffer, buffer_offset, pin_policy_amount_threshold, (short)((unit-1)*PIN_POLICY_AMOUNT_THRESHOLD_SIZE), PIN_POLICY_AMOUNT_THRESHOLD_SIZE);
 
-
         return (short)0;
     }
-
-// todo: amount policy: spending amount allowed before requiring pin
-
 
     /**
      * DEPRECATED - use exportAuthentikey() instead.
@@ -2366,13 +2418,14 @@ public class Satocash extends javacard.framework.Applet {
             ISOException.throwIt(SW_INVALID_PARAMETER);
 
         // generate a new ephemeral key
-        sc_ephemeralkey.clearKey(); //todo: simply generate new random S param instead?
-        Secp256k1.setCommonCurveParameters(sc_ephemeralkey);// keep public params!
-        randomData.generateData(recvBuffer, (short)0, BIP32_KEY_SIZE);
-        sc_ephemeralkey.setS(recvBuffer, (short)0, BIP32_KEY_SIZE); //random value first
+        if (ephemeral_privkey_transient) {
+            Secp256k1.setCommonCurveParameters(ephemeral_privkey);
+        }
+        randomData.generateData(recvBuffer, (short)0, SIZE_ECPRIVKEY);
+        ephemeral_privkey.setS(recvBuffer, (short)0, SIZE_ECPRIVKEY); //random value first
 
         // compute the shared secret...
-        keyAgreement.init(sc_ephemeralkey);        
+        keyAgreement.init(ephemeral_privkey);
         keyAgreement.generateSecret(buffer, ISO7816.OFFSET_CDATA, (short) 65, recvBuffer, (short)0); //pubkey in uncompressed form
         // derive sc_sessionkey & sc_mackey
         HmacSha160.computeHmacSha160(recvBuffer, (short)1, (short)32, CST_SC, (short)6, (short)6, recvBuffer, (short)33);
@@ -2386,7 +2439,7 @@ public class Satocash extends javacard.framework.Applet {
         // self signed ephemeral pubkey
         keyAgreement.generateSecret(Secp256k1.SECP256K1, Secp256k1.OFFSET_SECP256K1_G, (short) 65, buffer, (short)1); //pubkey in uncompressed form
         Util.setShort(buffer, (short)0, BIP32_KEY_SIZE);
-        sigECDSA.init(sc_ephemeralkey, Signature.MODE_SIGN);
+        sigECDSA.init(ephemeral_privkey, Signature.MODE_SIGN);
         short sign_size= sigECDSA.sign(buffer, (short)0, (short)(BIP32_KEY_SIZE+2), buffer, (short)(BIP32_KEY_SIZE+4));
         Util.setShort(buffer, (short)(BIP32_KEY_SIZE+2), sign_size);
 
@@ -2628,7 +2681,48 @@ public class Satocash extends javacard.framework.Applet {
                 return (short)0; 
         }
     }
-    
+
+    /**
+     * This function import the ECDSA secp256k1 NDEF authentikey private key during personalization.
+     * This private key is used to authenticate NDEF records.
+     * For privacy reason, this key is shared by multiple devices.
+     *
+     *
+     *  ins: 0xB
+     *  p1: rfu
+     *  p2: rfu
+     *  data: [ privkey (32b) ]
+     *  return: [none]
+     */
+    private short import_PKI_ndef_authentikey(APDU apdu, byte[] buffer) {
+
+        if (personalizationDone)
+            ISOException.throwIt(SW_PKI_ALREADY_LOCKED);
+
+        short bytesLeft = Util.makeShort((byte) 0x00, buffer[ISO7816.OFFSET_LC]);
+        if (bytesLeft < (short)32)
+            ISOException.throwIt(ISO7816.SW_WRONG_LENGTH);
+
+        // overwrite authentikey with privkey
+        // for Satocash, we can use the same authentikey for multiple cards, for better privacy
+        authentikey_private.setS(buffer, ISO7816.OFFSET_CDATA, SIZE_ECPRIVKEY);
+
+        // recover (uncompressed) pubkey
+        short offset = (short)0;
+        keyAgreement.init(authentikey_private);
+        keyAgreement.generateSecret(Secp256k1.SECP256K1, Secp256k1.OFFSET_SECP256K1_G, (short) 65, buffer, offset);
+        authentikey_public.setW(buffer, offset, (short)65);
+
+//        // compress pubkey
+//        if (buffer[(short)(offset+64)]%2 == 0){
+//            buffer[offset] = (byte)0x02;
+//        } else {
+//            buffer[offset] = (byte)0x03;
+//        }
+
+        return (short)0;
+    }
+
     /**
      * This function locks the PKI config.
      * Once it is locked, it is not possible to modify private key, certificate or allowed_card_AID.